refactor: 하드코드된 값을 분리하고 검증 함수들을 정적 함수화함
