+++ src/main/kotlin/racingcar/misc/Values.kt
@@ -1,0 +1,12 @@
+package racingcar.misc
+
+class Values {
+    companion object {
+        const val WIN_NUMBER = 4
+        const val MIN_CAR_NAME_LENGTH = 5
+        const val MIN_RANDOM_NUMBER = 0
+        const val MAX_RANDOM_NUMBER = 9
+        const val MIN_CAR_COUNT = 2
+        const val MIN_ATTEMPT_COUNT = 1
+    }
+}
