+++ src/main/kotlin/racingcar/view/InputView.kt
@@ -3,18 +3,16 @@ package racingcar.view
 import racingcar.exceptions.InputViewExceptions
 
 class InputView {
-    private val exceptions = InputViewExceptions()
-
     fun carNames(): List<String> {
         val input = readln()
         val names = input.split(",")
-        exceptions.validateNames(names)
+        InputViewExceptions.validateNames(names)
         return names
     }
 
     fun attemptCount(): Int {
         val input = readln()
-        exceptions.validateAttemptCount(input)
+        InputViewExceptions.validateAttemptCount(input)
         return input.toInt()
     }
 }
